+++ ScreenCapture/Shape/ShapeBase.h
@@ -8,10 +8,10 @@ public:
 	~ShapeBase();
 	virtual void paint() = 0;
 	virtual void paintDragger() {};
-	virtual void mouseMove(const int& x, const int& y) { };
-	virtual void mouseDrag(const int& x, const int& y, const UINT_PTR& modifiers) {};
-	virtual void mouseDown(const int& x, const int& y) { };
-	virtual void mouseUp(const int& x, const int& y) { };
+	virtual void mouseMove(const float& x, const float& y) { };
+	virtual void mouseDrag(const float& x, const float& y, const UINT_PTR& modifiers) {};
+	virtual void mouseDown(const float& x, const float& y) { };
+	virtual void mouseUp(const float& x, const float& y) { };
 	virtual void setCursor() {};
 public:
 	WinPin* win;
