+++ ScreenCapture/Shape/ShapeRect.h
@@ -8,10 +8,10 @@ public:
 	~ShapeRect();
 	void paint() override;
 	void paintDragger() override;
-	void mouseDrag(const int& x, const int& y,const UINT_PTR& modifiers) override;
-	void mouseDown(const int& x, const int& y) override;
-	void mouseUp(const int& x, const int& y) override;
-	void mouseMove(const int& x, const int& y) override;
+	void mouseDrag(const float& x, const float& y,const UINT_PTR& modifiers) override;
+	void mouseDown(const float& x, const float& y) override;
+	void mouseUp(const float& x, const float& y) override;
+	void mouseMove(const float& x, const float& y) override;
 	void setCursor() override;
 public:
 private:
